NOJIRA-0: e2e: object -> log -> text -> log -> object
Still need to figure out how to capture the full history of
changes, or for that matter how to change the object at all.
What would be nice would be an easy way to aop/proxy our way
to glory here. But I'm not even sure yet what the api should
look like.

Change-Id: I56cf1869c0bf53432138071eb8d8ae8b2722b16e
Reviewed-on: https://gerrit.gilt.com/6883
Tested-by: Eric Bowman <ebowman@gilt.com>
Reviewed-by: Eric Bowman <ebowman@gilt.com>