NOJIRA-0: basic time machine support
See the "time machine" test in OpmTest for how to use it.

Now our MapProxy instances hold on to their class, and to their
creation timestamps, so we can look through history.

Now objects must mix in OpmObject, which provides a method "magic"
which can be used by the implementation to recover the MapProxy.
Trying to figure out another way, but maybe it's not the end of
the world to have to mixin OpmObject. I guess an alternative would
be to track all objects, but that seems wasteful.

Also fixed a bug that wouldn't allow setting natives, like long.
That required some fancy footwork in the proxy implementation,
in particular looking specifically for the native java class names,
like "long" and "int".

"equals" & hash code now work reasonably.

Change-Id: Id9aede660c4f9d2f532e3511055cf518d33eb9b1
Reviewed-on: https://gerrit.gilt.com/7052
Reviewed-by: Eric Bowman <ebowman@gilt.com>
Tested-by: Eric Bowman <ebowman@gilt.com>