OPIUM-3: make key intrinsic so we can store nested OpmObjects
If you have an OpmObject that is composed of other OpmObjects, it's
pretty nice if you can persist that without thinking too hard about it.
Unfortunately to do that, we need the "timeline key" to be an intrinsic
part of the object, not something you supply extrinsicly when you create
a new timeline record. So this change is so that you have to supply a
key when you create a new instance, and then the runtime will be able to
automagically create or update any composed OpmObjects magically for
you.

I spent quite a long time trying to make the key type generic.
Unfortunately that completely screwed up the implicit RichOpmObject
conversion. So we compromised to make the key a String. Not ideal, and
maybe a second pass can fix that.

Perhaps somewhat annoying that you have to supply a key to instance; you
can see the mess in the non-mongo unit tests, where we pass "" as the
key all over the place. It might be reasonable to think about how we
could avoid that, although in practice I think these objects will
basically always be persisted to Mongo or something else, so it's
probably fine.

Change-Id: If006d2ec9896b44be4aa6b55e6c8bea7d00d9520
Reviewed-on: https://gerrit.gilt.com/9059
Reviewed-by: Eric Bowman <ebowman@gilt.com>
Tested-by: Eric Bowman <ebowman@gilt.com>