OPM-9: Now you can overwrite on put with an unpersisted instance
Before this fix, this test failed:

  test("overwrite succeeds") {
    import MongoTest.SimpleDomain
    val d1 =
      OpmFactory.instance[SimpleDomain]("sd1").
        set(_.name).to("d1").
        prune

    val d2 =
      OpmFactory.instance[SimpleDomain]("sd1").
        set(_.name).to("d2").
        prune

    put(d1)
    put(d2)

    val d3 = get[SimpleDomain]("sd1")
    assert(d3.map(_.name) === Some("d2"))
  }

...d3.name would come back as d1, because the update code did not handle
the case where the histories were not connected. Now we check for this
case: when we update, we actually load the old object and look for the
point where the histories diverge. If they never were actually
converged, then we delete the old and write the new. Otherwise we
stitch.

Change-Id: I7b2316867e2cc01c84c75715dd5fdcee6b14568d
Reviewed-on: https://gerrit.gilt.com/10828
Reviewed-by: Andrew Duffy <aduffy@gilt.com>
Reviewed-by: Miguel Fajardo <mfajardo@gilt.com>
Tested-by: Miguel Fajardo <mfajardo@gilt.com>