OPM-15: concurrent write support
To implement concurrent write support, we've added a distributed lock
per record to Mongo (using Mongo). Before modifying Mongo, OPM tries to
take a lock on the logical record (which may span many Mongo records).
Once it has that lock, it loads what is there, and determines if what is
trying to be saved is a replacement for (meaning, there is no shared
history), a fast-forward of (in Git terms), or a divergent history. The
first two cases succeed, and the last case will throw a
ConcurrentModificationException.

There is also a LockReaper mechanism, which is probably best coupled to
Leader Election, which will run a periodic job looking for locks which
are older than some timestamp, which gets automatically deleted. This is
a sort of garbage collection to remove locks which got abandoned,
perhaps due to a JVM restart. Each instance tracks the locks it created,
and tries to clean them up via a shutdown hook.

Also note that you can also specify the WriteConcern you want by
overriding the writeConcern method on LockManager. The default is SAFE.

Change-Id: I34638023f3a35f4a7c2922c075835a704a505ddf
Reviewed-on: https://gerrit.gilt.com/13389
Tested-by: jenkins bot <jenkins@hudson.qa.corp.gilt.com>
Reviewed-by: Eric Bowman <ebowman@gilt.com>