OPM-22: Much richer & more robust collection support.
There were a bunch of edge cases around collections that didn't work
right. Adam ran into one around Sets and had a partial fix; this is a
much more robust fix and includes a bunch of tests around it. In
particular a collection of collections of opm objects works now (see
NestedObjCollecionTest and ComplexCollectionTest).

Should be completely backwards compatible with previously-written opm
databases, though what is written by this change may not read correctly
by older versions (though in that case the older version was dealing
with a nasty bug anyhow).

Change-Id: I6927da3f0e75db764fefba27db7ce920e6de3d38
Reviewed-on: https://gerrit.gilt.com/14370
Tested-by: jenkins bot <jenkins@hudson.qa.corp.gilt.com>
Reviewed-by: Andrew Duffy <aduffy@gilt.com>
Reviewed-by: Eric Bowman <ebowman@gilt.com>
Tested-by: Eric Bowman <ebowman@gilt.com>