OPM-29: Adding search by 'contains' and 'in'
We needed some search functions to search within
arrays - both as fields in the datastore and as
the input of the search option. In this context,
if you want to search that an array field contains
a given value, use 'contains'. If you have a non-
array field and would like to see if it is one
of a list of values (essentially a chained 'or',
use 'in'. Example:

  case class Foo(name: String, tags: Set[String])

  search(_.name).in(Set("name1", "name2", "name3"))
  search(_.tags).contains("tag1")

Change-Id: I4768dad66357f5cc5cd47bbde6f939fd52c3314a
Reviewed-on: https://gerrit.gilt.com/17325
Tested-by: jenkins bot <jenkins@hudson.qa.corp.gilt.com>
Reviewed-by: Ryan Martin <rmartin@gilt.com>