Support backwards-compatibility of stored objects with None
When adding a new property to a new OPM object, existing
stored objects will be missing that property. If that new
property is an Option for backwards-compatibility, it is
implied that if it is missing from the stored OPM object
that it is a None. This use case was previously throwing
a NoSuchElementException, so avoid that and return None
instead.

Change-Id: I7a1fe90729a14cc45e820897e723e6c2f09dbb66
Reviewed-on: https://gerrit.gilt.com/38395
Reviewed-by: Myyk Seok <mseok@gilt.com>
Tested-by: jenkins bot <jenkins@hudson.qa.corp.gilt.com>
Reviewed-by: Eric Bowman <ebowman@gilt.com>